refactor(rpc): Move `RpcInvalidTransactionError` to `reth-eth-rpc-types`
Extend `CallFeesError` with similar variants that were used from the error
