+++ crates/rpc/rpc-eth-api/src/lib.rs
@@ -31,8 +31,8 @@ pub use reth_rpc_eth_types::error::{
     AsEthApiError, FromEthApiError, FromEvmError, IntoEthApiError,
 };
 pub use reth_rpc_types_compat::{
-    error, try_into_op_tx_info, CallFeesError, EthTxEnvError, IntoRpcTx, RpcConverter,
-    TransactionCompat, TransactionConversionError, TryIntoSimTx, TxInfoMapper,
+    try_into_op_tx_info, CallFeesError, EthTxEnvError, IntoRpcTx, RpcConverter, TransactionCompat,
+    TransactionConversionError, TryIntoSimTx, TxInfoMapper,
 };
 pub use types::{EthApiTypes, FullEthApiTypes, RpcBlock, RpcHeader, RpcReceipt, RpcTransaction};
