+++ crates/rpc/rpc-eth-types/Cargo.toml
@@ -32,6 +32,7 @@ reth-trie.workspace = true
 alloy-eips.workspace = true
 alloy-primitives.workspace = true
 alloy-consensus.workspace = true
+alloy-sol-types.workspace = true
 alloy-rpc-types-eth.workspace = true
 revm.workspace = true
 revm-inspectors.workspace = true
@@ -58,7 +59,6 @@ tracing.workspace = true
 itertools.workspace = true
 
 [dev-dependencies]
-alloy-sol-types.workspace = true
 serde_json.workspace = true
 
 [features]
