+++ crates/rpc/rpc-types-compat/src/lib.rs
@@ -11,7 +11,6 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
 pub mod block;
-pub mod error;
 mod fees;
 pub mod transaction;
